primitives for input/output dtype conversion
