+++ src/lightning_lite/plugins/precision/double.py
@@ -15,6 +15,7 @@ from contextlib import contextmanager
 from typing import Generator
 
 import torch
+from torch import Tensor
 
 from lightning_lite.plugins.precision import Precision
 
@@ -34,3 +35,6 @@ class DoublePrecision(Precision):
         torch.set_default_dtype(torch.float64)
         yield
         torch.set_default_dtype(default_dtype)
+
+    def convert_input(self, data: Tensor) -> Tensor:
+        return data.to(torch.float64) if torch.is_floating_point(data) else data
